Add BUILD.gn DEP validation
This CL adds a script that takes a directory and will validate that all
JavaScript dependencies that are imported are also defined in BUILD.gn,
and vice-versa.

=> node scripts/deps/sync-build-gn-imports.js\
    --directory=front_end/models/trace

The script is non recursive.

You can also execute the tests: `./node_modules/.bin/mocha
scripts/deps/tests`. These are only run locally and not yet run on CQ,
although I suspect we should have them run automatically.

This CL doesn't enable this script to run in anyway; the long term goal
is to have it run in a PRESUBMIT, however testing it locally has
uncovered quite a few invalid BUILD.gn files, so we'll need to fix them
one by one and enable the script incrementally.

Bug: 1402989
Change-Id: I587ba1a18a2edf7a126bda651cab79862e9ba525