Add the new "Commit" devtools timeline event (front-end side)
"Commit" is the step of the rendering pipeline where we send (also
known as "commit") layers from the main thread to the compositor
thread [1]. This patch adds a new performance timeline event for
"Commit" that will replace the existing "Composite Layers" step. For
backwards compatibility, the existing "Composite Layers" entry is not
removed.

This patch implements the frontend code and will be followed up with a
chromium-side patch to emit the renamed event.

[1] https://developer.chrome.com/articles/renderingng-architecture

Bug: 1191032
Change-Id: Ib8848178b97b91ae92806cf45c0eedb7cff767d5