[TextEditor] Reduce visual clutter due to Code Folding markers.
When switching to CM6 the Code Folding markers in the text editor
received additional border and background, which draws a lot of
attention to these markers, when it reality Code Folding should
actually help to not focus on the folded parts.

This restores the original CodeMirror5 styling of not using a
background and border, which is also consistent with how VSCode
renders these markers.

Before: https://imgur.com/A4FyMil.png
After: https://imgur.com/e4jaII6.png
Fixed: chromium:1403164
Bug: chromium:1371585
Change-Id: I31729b5e36184310f32a61846e64d6a577882bcb