[settings] Put 'searchAsYouType' setting into the correct spot.
With https://crrev.com/c/3865397 a new 'searchAsYouType' setting was
introduced. However since it was added to inspector_main, the setting
was only registered when debugging using the inspector app (basically
when debugging a Blink renderer), but the setting was checked in the
SearchableView which is also used in other apps (like the node_app),
and that functionality was thereby broken. This fixes the regression
by registering the setting in the appropriate spot (in main).

Bug: chromium:1344526, chromium:1371585
Change-Id: If5abae619e82f892fcbd05a3912649803c9f9a5b
Fixed: chromium:1403171