[cleanup] Stop lying about the type of the SourceMapV3 object.
This simplifies and aligns the `SourceMapV3` type with the
specification[^1], and updates tests and consuming code accordingly. The
main point here is that most of the fields of a sourcemap JSON object
are optional, and there are no guarantees that the `sources` or the the
`sourceRoot` fields contain valid URL strings.

[^1]: https://docs.google.com/document/d/1U1RGAehQwRypUTovF1KRlpiOFze0b-_2gc6fAH0KY0k

Bug: chromium:1403432
Change-Id: Ia66823632268d69a4a55f07c88193e512e712acd