[bindings] Fix handling of conflicting source maps.
This basically adds proper support for "hot module replacement" to
the `CompilerScriptMapping` by resolving conflicts between source maps
within the same target using the following strategy:

1. If two source maps from different scripts reference the same authored
   file, and neither of the source maps provides content for the file
   and agree on the ignore-list hint, we consider this to be the same
   authored file.
2. If two source maps from different scripts reference the same authored
   file, and both source maps carry the same `sourcesContent` for the
   file and agree on the ignore-list hint, we consider this to be the
   same authored file.
3. In all other cases we assume that there is a conflict, and we further
   assume that the intention here is to replace the authored file with a
   newer version (via "hot module replacement"). We therefore unlink
   the previous `UISourceCode` for the authored file and mark all
   mappings from earlier source maps to this `UISourceCode` as stale,
   meaning that future attempts to resolve the mapped locations from the
   associated scripts will not find a `UISourceCode` coming from the
   `CompilerScriptMapping` (but will fall back to the default mappings).

The "stale source map" case is especially interesting, since we need to
ensure that not only UI to raw location lookups will ignore stale
UISourceCodes, but also when mapping raw to UI locations, we need to
double check that the UISourceCode we find is still active for the
source map associated with the script.

Drive-by-fix: Add documentation and more test cases for this
functionality.

Doc: http://go/devtools-source-identities
Fixed: chromium:1403432
Bug: chromium:1403362
Change-Id: Ic54043e4c6d6ca703bdcec6d300e282518503c41