[e2e] Fix: Remove aync describe in can-create-panels_tests.ts
Passing an aync callback to describe can cause unintended behavior in mocha helpers.

Remove the `async` keyword to the describe callback. It is not needed.

If in the future async work is needed before a test is run passing an async callback to `before` is the recommended approach.
See https://mochajs.org/#using-async-await

Bug: None
Change-Id: I6ad2cef69ece58a91d333932d35e8ed5c6c1436b