Change process.url to be a string in RendererHandler
The RendererHandler represents process URLs as a `URL` object, but URL
objects cannot be sent via a `postMessage` from a worker, which prevents
us from moving the parsing into a worker.

Therefore, we'll instead represent the URL as a string. It's not quite
so nice, but any consumers can easily `new URL` it once it gets back
from the worker onto the main thread.

In the RendererHandler we still call `new URL` to ensure a URL is valid,
otherwise the handler starts to keep processes around with invalid URLs.

Bug: 1399473
Change-Id: Ic689ae6154cfc246fea0861c79374fccc6f4da3a