Fix hit testing for Elements tree items
A previous fix to this issue couldn't preserve the original behaviour,
and hit testing could not find the element on that row if the cursor
is above the indent space.

With this CL, hovering/selecting anywhere on the row (including indented
space) would highlight/select the element on that row, whether the
overlay scrollbar is enabled/disabled.

We will use '::after' psuedo element on the TreeElement to overlay the
available indented space, such that hovering/selecting on indentation
will now hit the psuedo element instead of its ancestor's padding area.

Fixed: 1404874
Change-Id: I31174dca4e37826d696a82b1f4c4601f9cef28e5