[cleanup] Fix the `ProjectStore.uiSourceCodes()` foot-gun.
Previously `ProjectStore.uiSourceCodes()` was implemented to return an
internal array of `UISourceCode`s that was mutated as sources got added
and removed from the project. This led to subtle bugs when removing
sources from a project by just iterating over this array, where because
of the funny compaction logic the iterator would skip over items and
miss them.

The internal implementation of this is also unnecessary complex by using
both a Map and an Array. We can drastically simplify this by simply only
using the Map and having `ProjectStore.uiSourceCodes()` return an
iterable that adheres to the `Map#values()` rules, and in particular is
safe to use while removing items from the store.

Bug: none
Change-Id: I751126906357c5e27f08c3bf7a6e278094d801d1