Refresh breakpoints for live edit only if the source has changed
When a binding is created, the NetworkPersistenceManager will
make sure to synchronize the already fetched content of a network
resource with its override version. As a result, the script
source content is re-set, and all breakpoints that were
previously set on the network resource are refreshed
(i.e. removed and reset) in order to reflect the up-to-date state.

This update of breakpoints even happens if the two sources
do not differ. In this case, we unnecessarily remove and reset
breakpoints.

With instrumentation breakpoints, this logic would sometimes
remove the breakpoints that we've already set, just to reset
it at the exact same position.

In order to avoid removing breakpoints that we've already
correctly set, this CL makes sure to only update the
source code if the new source code differs from the old one.

This also adds a test for network persistence using the new
MockBackend.

Bug: 1396298, 1395337
Change-Id: I057cd6a49cec45ad88c837b602bcace9cd3a7968