[source] Ensure syntax tree is available for popover.
Previously we just assumed that the syntax tree is available under the
cursor position, but CodeMirror does the parsing and AST construction
lazily (and using idle callbacks), so there's no guarantee here. With
this CL we now give CodeMirror a chance to complete parsing at least up
to the cursor position, just like what we do in other places.

Fixed: chromium:1403198
Bug: chromium:1383892, chromium:1361862, chromium:1383451
Change-Id: If9cefc6384b34152fec2474f1426ee8777d93496