[cleanup] Remove initiator from `SourceMap`s.
The fact that sourcemaps know about initiators presents a layering
violation and hiding the logic to request source contents within
the `SourceMap` class makes it difficult to tackle the problem of
dealing with resources that fail to load.

Drive-by-fix: Also remove i18n from the source map error messages. We
currently translate only a few of the errors, which leads to a quite
inconsistent experience; it's better to not translate them at all for
now (especially since we have no control over the error text that comes
out of JSON.parse() for example).

Bug: chromium:1262113
Change-Id: I45df23b99e2fc3d42a47e3999df4ab8d75b54541