[cleanup] Replace Storage.Item class for breakpoints
This CL replaces the 'Storage.Item' class with a new
'BreakpointStorageState' interface thats directly used by the
'Breakpoint' class as a member.

The two advantages are 1), we don't need to keep the properties
between 'Breakpoint' and 'Storage.Item' in sync (including types).
And 2), changing what we store per breakpoint becomes easier.

R=jarin@chromium.org

Bug: 1027458
Change-Id: I10e523202193d0d77aa611cdfbbf516b1232fc95