Introduce fixed positioning containing block getter
Elements whose position is set to 'fixed' are positioned relative to the
initial container established by the viewport unless an ancestor
establishes a containing block. This CL adds a helper that gets an
element's nearest ancestor that establishes a fixed positioning
containing block.

This helper can aid in positioning an element relative to the viewport
even if the element is a descendant of an element that establishes a
containing block.

Bug: none
Change-Id: Ib24e5227b5aee5684f2efea9c057b3bfa75b7994