Render the perf panel in the component server
In order to be able to more easily work with the panel, and also to
write tests and in particular screenshot tests, we need to be able to
run it independently. This CL enables that and builds the first example.

We have to do a fair amount of setup to enable this and set up some
global options, I hope over time we can work to reduce these and make
the panel more self contained, but for now I don't mind the extra setup
cost in order to get the panel rendering and loading in a trace.

Bug: 1408406
Change-Id: I3fb5dcf817e4f7e4e90631df77541faef169d476