[breakpoint] Add 'isLogpoint' property to each stored breakpoint
The property itself is not set or used yet. It's only piped through
all the necessary methods.

'isLogpoint' is required as we want to be more flexible how we modify
the breakpoint condition before we send it to V8. It's necessary
that we store the condition as it was entered by the user and store
any additional info separately without encoding it in the condition.

Bug: 1027458
Change-Id: I1d261d95d4b8e24f6297d0d4b539ccb74117c5e3