[Performance] Correct JS frame lengths when parented by unknown v8 events
generateJSFrameEvents() turns JS samples into JSFrames (with start and end).
Some logic checks to see if we're inside an invocationEvent (aka
EvaluateScript or FunctionCall) before we create JSFrames. The allowlist
of invocation events is now dated, as V8 emits some new trace events.

Meanwhile, http://crrev.com/c/2443620 paved a path for sitautions when
no invocation parents are present. This path was untouched in this CL.

Bug: 1384182
Change-Id: Ib0c74c28f0c17dc5c8142aa812da9ccbebbc9fdd