[breakpoint] Store 'isLogpoint' for breakpoints
When this CL lands, the following condition will hold: `isLogpoint`
is true iff the breakpoint condition starts with the logpoint prefix
and ends with the logpoint suffix. Both for new breakpoints as well
as already stored breakpoints.

Going forward we can actually read the `isLogpoint` property instead
of checking the prefix/suffix. Once we updated all relevant sites,
we can store the breakpoint/logpoint condition as the user entered
it.

R=jarin@chromium.org

Bug: 1027458
Change-Id: Iae69bb114649f667ff33c286ba892d304ec86db3