[sourcemaps] Don't append the "? [sm]" suffix.
Prior to this change, whenever there was a clash between the URL of a
script with a sourcemap and one of the source files mentioned in that
sourcemap, DevTools would append a "? [sm]" suffix to the source URL
of that inline source. This was particularly visible (and annoying)
with modern development setups using Vite (for example the Vue and
Svelte development setups), where source files would always end up
having the "? [sm]" initially, as Vite was sending the generated code
under e.g. App.svelte referencing App.svelte as source.

Even worse, with Hot Module Replacement (HMR) in these modern setups,
the source file name would change when the file changes for the first
time, because then Vite sends App.svelte?ts=<TIMESTAMP> referencing
App.svelte as source, and DevTools no longer sees a conflict in names.

The consequences here were that no only we had (initially) confusing
names for the files, which changed over as the developer was editing
their source files, but also that breakpoints would be dropped on the
floor in doing so, or breakpoints set on the files after hot module
replacement would no longer be present after reloading (again because
the URL is different on fresh state).

This was apparently added as ducktape a long time ago to workaround the
problem that breakpoints were only keyed on URL, and would thus not work
correctly if source file and source mapped file had the same URL. Since
https://crrev.com/c/4177296 (issue 1408032) we properly key breakpoints
on both URL and resource type, so this workaround is no longer needed.

Doc: http://go/devtools-source-identities#bookmark=id.vgx3csupe59s
Fixed: chromium:1408033
Change-Id: Ibd386b2c94e7dca9b75f04c4b639ce65c21efae9