Add @vscode/web-custom-data package to front_end/third_party
This CL includes the @vscode/web-custom-data package. It's maintained
by MS for vscode and includes MDN documentation for various CSS and
other web features.

We want to display this info also in DevTools. Given that the file has
an uncompressed size of ~730kB, we need to fetch the
browsers.css-data.json file remotely, similar to locale files. As such
we don't include the file in our GRD.

For now, we DEPS the file in via the elements panel as this will be
the sole user for now. Should we need the information in other parts
of DevTools we might move the dependency to front_end/BUILD.gn.

DISABLE_THIRD_PARTY_CHECK=Already include the package in the build
output

R=bmeurer@chromium.org

Bug: 1401107
Change-Id: I104a09b5c49d2b811e31a853c95ba8c3034b2dee