[breakpoint] Rename 'preferLogpoint' to 'isLogpoint'
This CL renames the constructor argument for the breakpoint edit
dialog from 'preferLogpoint' to 'isLogpoint'.

We know up-front in all cases whether we want to have a logpoint or
a conditional breakpoint, so we shouldn't do any guessing in the
constructor.

Either we pass an existing breakpoint/logpoint for editing, or we
create a new breakpoint/logpoint via context menu. In both cases
we know what kind we want to create.

R=jarin@chromium.org

Bug: 1027458
Change-Id: I1e1427a97c07968187fe33a19381a53c5d6cb23c