Add 'Breakpoint#backendCondition()' and 'BackendCondition' type
We want to distinguish between the breakpoint condition as the user
entered it, and as we send it to V8. For logpoints we wrap the
condition in a console.log and we also plan to add 'sourceUrl'
comments in the future.

For this, we add a new branded type to also reflect this plan in the
type system. We also add a bottleneck on the `Breakpoint` class where
the modifications to the condition will happen.

For now, `condition()` and `backendCondition()` return the same
string as the migration is under way.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: Ia608f5c668ebad2ae12961b271a1cae06bad6212