[Colors] Account for powerless components
Some colors define powerless copmonents, e.g., when lightness is 100, hue and saturation are irrelevant in the hsl space. With this CL, powerless components are set to 0.

In order to simplify testing and the API a little, also stop returning the "as-authored" original text for Color's asString method. Callers explicitly interested in showing the as-authored value should call the getAuthoredText accessor now.

Bug: 1073895
Change-Id: I6a93fbd5af4b8ae007122360bea3acb67c4b6aa5