[e2e] Unflake HMR breakpoint migration test.
Breakpoint setting can take more time than loading the new content (and
is also unrelated to the new content loading), so we need to properly
wait for the breakpoint to migrate here. This is supported by the
findings that in the front-end screenshots of both failed runs mentioned
in https://crbug.com/1409857, the breakpoint is already set.

Bug: chromium:1408033, chromium:1336590
Change-Id: Ia5542cf72c192efab73a3aaf757a6a68c8af1d1d
Fixed: chromium:1409857