Store breakpoint conditions as entered by the user
This CL switches how we store logpoints. Instead of storing the
already wrapped (in console.log) expression, we store the expression
as it was entered by the user and wrap it on-demand.

To make this work nicely with existing logpoints, we add a settings
migration that removes the prefix/suffix.

R=jarin@chromium.org, kimanh@chromium.org

Bug: 1027458
Change-Id: I58a2f0192d5b96fbe1afcf30b2015cb1360067ea