Refactor the click helper
The click helper now only accepts selectors so that it is
abel to recover from unstable DOM situations. A new deprecated
helper clickElement is introduced for tests that cannot use a
selector currently. This should serve as a red flag that the
test might suffer from flakines.

Bounding box and offset computation is removed for now. New
Puppeteer version allows providing relative click coordinates
so we need to use that.

Bug: 1409768
Change-Id: I94a272184530842476bb2a879d862c4bd0930dba