[BreakpointView] Show differentiating path for ambiguous file names
For ambiguous file names, we now show a part of the path to that
file that helps to differentiate the breakpoint groups next to
the file name.

When differentiating ambiguous files, we try to find a part of the
url that is uniquely identifying a path in terms of the other
files, if possible. From that unique path, we only show the right
most folder name. If we cannot find such a unique path, we show the
whole path instead.

This happens e.g. for React, where it is common to have an index.js
file for each component.

Fixed: 1403924
Change-Id: I5d7033b38b30dd875bee5a09389a6ebe09643e12