[cleanup] Remove redundant ScriptPosition and LocationRange types.
These classes were basically 1:1 wrappers around the protocol types, and
there's no reason to have them as separate classes as well. It just
makes things unnecessarily challenging to understand (because it isn't
immediately obvious what the difference is) and harder to reason about.

Also rename the RawLocationRange to LocationRange now and use that type
in the appropriate places.

Bug: chromium:1409365
Change-Id: I9ce5a3c272c5592a4ee9322fcdbf07832bae7063