Enable roll_deps.py to use working tree of Chromium repository
Currently, roll_deps.py only supports to roll up ToT of Chromium
repository. If working tree is dirty, it fails in
`git checkout origin/main`.

This CL adds to it a --ref option to control which of ToT or worknig
tree is used. This enables a new development flow: Write CDP changes
and backend code, and check it using frontend before CL lands.

Bug: none
Change-Id: Ic386d62addec124e9c4c01c3a0bb4469e9193de9