[bindings] Introduce uiLocationRangeToRawLocationRanges().
This introduces a new uiLocationRangeToRawLocationRanges() method to the
mapping layer which is used by BreakpointManager.possibleBreakpoints()
to determine the set of raw location ranges within which to ask V8 for
breakable locations. Previously the BreakpointManager assumed that
contiguous UI location ranges would always map to contiguous raw
location ranges, but in the presence of inline scripts and source maps
that's clearly not true.

The new uiLocationRangeToLocationRanges() takes into account even
special cases like holes that can result from funny source maps, and in
the future should also be useful to compute the skip ranges for the
auto-stepping logic.

For the SourceFormatter ScriptMapping we kept the assumption that the
ranges must be contiguous even though they don't necessarily are (for
formatted documents with inline scripts), but the whole SourceFormatter
functionality is basically deprecated at this point and will soon be
removed (in-place pretty printing was enabled by default with M-111
already).

Drive-by-fix: End offsets for inline scripts are exclusive, so
we need to treat them like that for the purpose of the location
mappings (especially ResourceScriptMapping and ResourceMapping).

Bug: chromium:1409365
Change-Id: I3ec7db0ad32272a0af96c2c06f8506d4886bbe9e