Only reveal the location for removed breakpoints
This CL updates the `BreakLocation` revealer. Until now, the revealer
would do two things: 1) Jump to the location and file where the
breakpoint is, and 2) open the breakpoint edit dialog.

This makes sense as the only use-site is currently from the
breakpoint sidebar pane. Every breakpoint clicked there is not
removed.

With the upcoming changes to console messages that originate from
logpoints, this distinction becomes more important: We should
only open the breakpoint edit dialog when the breakpoint wasn't
removed yet. This is because a user may click on a message anchor
even after the breakpoint was removed.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: Ie588fcdf67162e8f8d0a77e4e44445c779d77f8d