Hide query params for files from source maps.
It is common for webpack to add a hash after each file in the
source map. These are not meaningful to the user and best left
hidden. Because files from source maps aren't the result of
http get queries it is reasonable to assume these are rarely
needed. The exception to the new rule is if the filename would
otherwise be empty we keep the query params.

Fixed: 1402657
Change-Id: I33649f37e65a7c44c3614124f6a4908132b286ca