Fix BreakLocation revealer sometimes not showing edit dialog
This CL attempts to fix occurrences where the breakpoint edit dialog
would not open even though it should. We suspect that this can happen
when the "Sources" panel was not loaded yet. In that case, loading the
sources panel and editor races with drawing the breakpoint edit
dialog.

This CL forwards the promise for showing the "Sources" panel to
callers so revealers can wait for it.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: If24e1d9bf4dd6d5733ba2d7ac9742a30792015da