[sourcemaps] Sort mappings only once when using the index map format.
Previously when using the index map format[^1], we'd sort all the
mappings that we had seen until that point after every single section,
which scales really poorly as more sections are combined within a
single source map.

With this patch we only sort once after all the mappings from all the
sections are processed. On the bundle10000.js.map example from the
referenced bug, the time to execute `#ensureMappingsProcessed()` goes
from ~22s to ~120ms, so roughly a 183x improvement.

[^1]: http://doc/1U1RGAehQwRypUTovF1KRlpiOFze0b-_2gc6fAH0KY0k#heading=h.535es3xeprgt

Fixed: chromium:1411617
Also-By: Tobias Koppers <tobias.koppers@googlemail.com>
Change-Id: I64d6347f37152f13c2b2abde6484c5540946a46f