[linkifier] Add option to reveal the breakpoint edit dialog
This CL adds a new configuration option when we linkify script
locations. If the location we want to linkify also has breakpoint
set, then we can optionally reveal the breakpoint edit dialog
(after also jumping to the corresponding source location).

This functionality will be used to polish the UX around logpoints and
conditional breakpoints.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: I593d250e5b270a4c278afe1665ee023a36e836a6