[bindings] Properly restore breakpoints even in case of code-splitting.
Currently there's no notion of "UISourceCode changed" event within
DevTools, so the only way that the BreakpointManager and other pieces of
the system get informed about changes to mappings is via
`UISourceCodeAdded` and `UISourceCodeRemoved` events. That means in
order to properly restore breakpoints in the light of code-splitting we
need to change the logic within the CompilerScriptMapping to always
recreate `UISourceCode`s in case of conflicts and copy the mappings of
the compatible source maps over.

Fixed: chromium:1412033
Bug: chromium:1409365
Change-Id: I875569bbbf782f7170cc4a7e86c5cc73404eedfb