Add class for lazy loading CSS property documentation
This CL adds a small wrapper class that is reponsible for fetching
the @vscode/web-custom-data CSS info.

Creating a new instance immediatly triggers the fetch and a small
helper method allows looking up info by CSS property name.

Note that we intentionally don't return any promise to the data
from the lookup method. We want to show the info on hover and if
we return a promise we have to explicity handle the case where the
promise takes too long to resolve.

R=ergunsh@chromium.org

Bug: 1401107
Change-Id: Ia09d87036ac9d0d28f328a3501f8cc452812b3b3