Fix error in button test
I noticed an error when running the button tests:

```
==== PASS: Button/toolbar button can be clicked
ERROR: 'ScheduledRender: rendering devtools-button:'

ERROR: Error: Toolbar button does not accept children
Error: Toolbar button does not accept children
    at #render (http://localhost:9876/base/out/Default/gen/front_end/ui/components/buttons
/Button.js:138:23)
    at wrappedCallback (http://localhost:9876/base/out/Default/gen/front_end/ui/components
/helpers/scheduled-render.js:31:32)
    at #handleWork (http://localhost:9876/base/out/Default/gen/front_end/ui/components/ren
der_coordinator/RenderCoordinator.js:132:34)
    at http://localhost:9876/base/out/Default/gen/front_end/ui/components/render_coordinat
or/RenderCoordinator.js:187:46
===============================================

SUCCESS: 1 passed (0 skipped)
[jacktfranklin src/devtools-frontend] $ npm run auto-unittest
```

This is because we set inner text in the tests, but we do that for all
variants, even though some button variants do not support text.

The fix is to only set innerText if the button variant is appropriate,
and as a drive by I also fixed the typo in the error message.

Bug: none
Change-Id: I1b180dd56583134af5f667dbab6012bb78f388e0