Display MDN docs by hovering over CSS properties in DevTools
This CL adds a small lit component that displays the MDN docs
for a CSS property. The info is fetched remotely similar to
locale files.

Note that the TypeScript type declaration is currently duplicated
in "WebCustomData.ts" and "CSSPropertyDocsView.ts". This is to
prevent a cyclic dependency between "elements/components" and
"elements".

Screenshot: https://i.imgur.com/apOEOp8.png

Fixed: 1401107
Change-Id: I5afd79288806c4220caa3b2b1876fbe4e01110f7